Reverted most of changes from the latest commits. Please please please do not commit such stuff again, avoiding compiler warnings must not be a priority and the code that produces no warning is not always correct.

+++ src/interfaces/gtk/ec_gtk_targets.c
@@ -415,7 +415,7 @@ static void gtkui_delete_targets(GtkWidget *widget, gpointer data) {
    GtkTreeModel *model;
    struct ip_list *il = NULL;
 
-   switch((size_t)data) {
+   switch((int)data) {
       case 1:
          DEBUG_MSG("gtkui_delete_target: list 1");
          model = GTK_TREE_MODEL (liststore1);
